Split analytic ray tracer into dedicated AnalyticRayTracing module
Moves CPPAnalyticRayTracing into NuRadioMC/SignalProp/AnalyticRayTracing/
and trims the renamed single_layer_analytic_raytracer.py down to the
ray_tracing_2D implementation and its standalone helper functions,
recreating analyticraytracing.py as the thin ray_tracing wrapper class.
Also extracts the birefringence math helpers into
NuRadioMC/utilities/birefringence.py and caches the binary birefringence
ice-model data via functools.lru_cache.

Re-applies the same restructuring as PR #1087 (refactor/ray-tracer) on top
of this branch's diverged analyticraytracing.py, so this branch can be
merged instead of that one.
