+++ .github/workflows/deploy_documentation.yaml
@@ -31,7 +31,7 @@ jobs:
           pip install -e .[dev,proposal,cr_interpolator]
           export PYTHONPATH=$PWD:$PYTHONPATH
           echo $PYTHONPATH
-          NuRadioMC/SignalProp/install.sh
+          NuRadioMC/SignalProp/AnalyticRayTracing/install.sh
 
       - name: Debugging
         run: |
