+++ NuRadioMC/utilities/medium.py
@@ -11,6 +11,8 @@ import numpy as np
 import os
 from NuRadioReco.utilities import units
 
+import functools
+
 logger = medium_base.logger
 
 try:
@@ -103,9 +105,14 @@ class ARA_2022(medium_base.IceModelSimple):
 class birefringence_medium(medium_base.IceModelBirefringence):
 
     def __init__(self, bir_model='southpole_A'):
+        f = self._load_binary_data(bir_model)
+        self.load_birefringence_model(bir_model=f)
+
+    @functools.lru_cache(maxsize=8)
+    def _load_binary_data(self, bir_model):
         # from https://link.springer.com/article/10.1140/epjc/s10052-023-11238-y
         filepath = os.path.dirname(os.path.realpath(__file__)) + '/birefringence_models/birefringence_' + bir_model + '.npy'
-        super().__init__(bir_model=np.load(filepath, allow_pickle=True))
+        return np.load(filepath, allow_pickle=True)
 
 
 class mooresbay_simple(medium_base.IceModelSimple):
@@ -295,7 +302,7 @@ class greenland_perturbation(greenland_firn):
     def __init__(self):
         greenland_firn.__init__(self)
 
-    def _compute_default_ice_model_radiopropa(self,discontinuity=False):
+    def _compute_default_ice_model_radiopropa(self, discontinuity=False):
         """
         Computes a default object holding the radiopropa scalarfield and necessary radiopropa
         moduldes that define the medium in radiopropa. It uses the parameters of the medium
